requirements.txt
numpy
flask
pymysql
python-dotenv
scikit-learn
xgboost
matplotlib
pandas
google-generativeai
lightgbm
